pyproject.toml
[project]
name = "aperture"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "gitpython>=3.1.45",
    "maya-stubs>=0.4.2",
]

[tool.ruff]
line-length = 100

[tool.pyright]
typeCheckingMode = "off"
extraPaths = [
    "./.stubs/",
    "./.venv/lib/python3.11/site-packages",
    "/usr/autodesk/maya/lib/python3.11/",
    "/usr/autodesk/maya/lib/python3.11/site-packages",
    "/usr/autodesk/ApplicationPlugins/ngskintools2/Contents/scripts",

]

[dependency-groups]
dev = [
    "debugpy>=1.8.17",
    "maya-stubs>=0.4.2",
    "pyside6>=6.10.0",
    "ruff>=0.14.4",
]

